Fix isolated Telegram active-run ingress
