+++ docs/channels/telegram.md
@@ -293,7 +293,7 @@ curl "https://api.telegram.org/bot<bot_token>/getUpdates"
 - Inbound messages normalize into the shared channel envelope with reply metadata, media placeholders, and persisted reply-chain context for Telegram replies the gateway has observed.
 - Group sessions are isolated by group ID. Forum topics append `:topic:<threadId>` to keep topics isolated.
 - DM messages can carry `message_thread_id`; OpenClaw preserves it for replies. DM topic sessions split only when Telegram `getMe` reports `has_topics_enabled: true` for the bot; otherwise DMs stay on the flat session.
-- Long polling uses grammY runner with per-chat/per-thread sequencing. While a text-only turn is already in flight, later plain-text follow-ups in the same account/chat/thread use a narrow active-run ingress lane so they can reach the queue/steer policy without waiting for the running handler. Media and media-group updates remain on the normal serialized lane. Overall runner sink concurrency uses `agents.defaults.maxConcurrent`.
+- Long polling uses grammY runner with per-chat/per-thread sequencing. While a text-only turn is already in flight, later plain-text follow-ups in the same account/chat/thread use a narrow active-run ingress lane, including the isolated-spool polling path, so they can reach the queue/steer policy without waiting for the running handler. Media and media-group updates remain on the normal serialized lane. Overall runner sink concurrency uses `agents.defaults.maxConcurrent`.
 - Multi-account startup bounds concurrent Telegram `getMe` probes so large bot fleets do not fan out every account probe at once.
 - Long polling is guarded inside each gateway process so only one active poller can use a bot token at a time. If you still see `getUpdates` 409 conflicts, another OpenClaw gateway, script, or external poller is likely using the same token.
 - Long-polling watchdog restarts trigger after 120 seconds without completed `getUpdates` liveness by default. Increase `channels.telegram.pollingStallThresholdMs` only if your deployment still sees false polling-stall restarts during long-running work. The value is in milliseconds and is allowed from `30000` to `600000`; per-account overrides are supported.
